Added outline for models and made possible to resize, move and rotate them comfortly

+++ Camera/Camera.cpp
@@ -13,7 +13,7 @@ void Camera::updateMatrix(float FOVdeg, float nearPlane, float farPlane){
     glm::mat4 projection = glm::mat4(1.0f);
 
     view = glm::lookAt(Position, Position + Orientation, Up);
-    projection = glm::perspective(glm::radians(FOVdeg),(float)(width/height),nearPlane, farPlane);
+    projection = glm::perspective(glm::radians(FOVdeg),static_cast<float>(width)/static_cast<float>(height),nearPlane, farPlane);
 
     cameraMatrix = projection * view;
 }
